requirements.txt
streamlit
openai
faiss-cpu
numpy
gspread
oauth2client
pandas